+++ .gitignore
@@ -7,3 +7,5 @@
 *.xml
 *.xlsx
 *.log
+*.old
+*.pid
