+++ ToolboxClient.py
@@ -13,12 +13,18 @@ import json
 import socket
 
 
+def get_hosts():
+    hostlist = []
+    with open("hosts.txt", "r") as hosts:
+        host = hosts.readline().rstrip()
+        hostlist.append(host)
+    return hostlist
+
+
 def post_iof3_resultList(event, file):
-    hosts = open("confighosts.txt")
-    while True:
+    hosts = get_hosts()
+    for host in hosts:
         try:
-            host = hosts.readline().rstrip()
-            if not host: break
             print("sending results to " + host)
 
             # post sport software IOFv3 results file
@@ -30,24 +36,20 @@ def post_iof3_resultList(event, file):
             print r.status_code, r.text
         except:
             print("failed to send results to " + host)
-    hosts.close()
 
 def put_clubcodetable(file):
-    hosts = open("confighosts.txt")
-    while True:
-        host = hosts.readline().rstrip()
-        if not host: break
+    hosts = get_hosts()
+    for host in hosts:
         print("sending clubs to " + host)
         
         url = host + '/api/clubs'
         f = {'file': open(file, 'r')}
         r = requests.put(url, files=f)
         print r.status_code, r.text
-    hosts.close()
-    
+
 def put_cclasstable(event, file):
-    hosts = open("confighosts.txt")
-    while True:
+    hosts = get_hosts()
+    for host in hosts:
         host = hosts.readline().rstrip()
         if not host: break
         print("sending classes to " + host)
@@ -56,11 +58,10 @@ def put_cclasstable(event, file):
         f = {'file': open(file, 'r')}
         r = requests.put(url, files=f)
         print r.status_code, r.text
-    hosts.close()
-    
+
 def put_eventtable(file):
-    hosts = open("confighosts.txt")
-    while True:
+    hosts = get_hosts()
+    for host in hosts:
         host = hosts.readline().rstrip()
         if not host: break
         print("sending events to " + host)
@@ -69,19 +70,16 @@ def put_eventtable(file):
         f = {'file': open(file, 'r')}
         r = requests.put(url, files=f)
         print r.status_code, r.text
-    hosts.close()
-    
+
 def put_entrylist(event, file):
-    with open("confighosts.txt", "r") as hosts:
-        while True:
-            host = hosts.readline().rstrip()
-            if not host: break
-            print("sending entries to " + host)
-            
-            url = host + '/api/event/' + event + '/entries'
-            f = {'file': open(file, 'r')}
-            r = requests.put(url, files=f)
-    
+    hosts = get_hosts()
+    for host in hosts:
+        print("sending entries to " + host)
+        
+        url = host + '/api/event/' + event + '/entries'
+        f = {'file': open(file, 'r')}
+        r = requests.put(url, files=f)
+
 def poll_for_results(eventcode, dir):
     lastUpdate = None
     while True:
@@ -110,7 +108,7 @@ def relay_wibox():
     #    for line in f:
     #        hosts.append(line.rstrip())
 
-    hosts = ['http://localhost']
+    hosts = get_hosts()
     while True:
         punches = reader.poll_punch()
         if len(punches) > 0:
@@ -124,7 +122,6 @@ def relay_wibox():
                     r = requests.post(url, headers=header, data=json.dumps(punch))
                     print 'Sent {} to {}'.format(punch, h)
     return
-    
 
 if __name__ == '__main__':
     method = sys.argv[1]
@@ -135,21 +132,21 @@ if __name__ == '__main__':
         python ToolboxClient.py iof3 2016-02-27-1 testdata.xml 
         '''
         post_iof3_resultList(sys.argv[2], sys.argv[3])
-    
+
     elif method == 'monitor':
         ''' 
         monitor a folder for new results and post them
         usage: python ToolboxClient.py monitor 2016-05-05-1 ./results_go_here 
         '''
         poll_for_results(sys.argv[2], sys.argv[3])
-        
+
     elif method == 'wibox-relay':
         '''
         relay data incoming from a given wi-box on TCP to the host via a POST request
         usage: python ToolboxClient.py wibox-relay
         '''
         relay_wibox()
-        
+
     elif method == 'telemetry':
         ''' 
         mimic a call coming in from a remote punch box
@@ -164,35 +161,35 @@ if __name__ == '__main__':
         header = {'content-type': 'text/json'}
         r = requests.post(url, headers=header, data=json.dumps(d))
         print r.text
-    
+
     elif method == 'clubs':
         '''
         post a json file containing the mapping of club code to full name
         python ToolboxClient.py clubs clubcodes.json
         '''
         put_clubcodetable(sys.argv[2])
-        
+
     elif method == 'classes':
         '''
         post a json file containing the mapping of class code to full name
         python ToolboxClient.py classes 2016-02-27-1 classcodes.csv
         '''
         put_cclasstable(sys.argv[2], sys.argv[3])
-        
+
     elif method == 'entries':
         '''
         post a xml file containing the entries for the meet
         python ToolboxClient.py entries 2016-02-27-1 meetentries.xml
         '''
         put_entrylist(sys.argv[2], sys.argv[3])
-        
+
     elif method == 'events':
         '''
         post a tsv file containing the event information
         python ToolboxClient.py events eventinfo.tsv
         '''
         put_eventtable(sys.argv[2])
-        
+
     # elif method == 'prep-db':
         # '''
         # prep the database with classes, teams, and anything else...
@@ -200,4 +197,3 @@ if __name__ == '__main__':
         # '''
         # put_clubcodetable('clubcodes.json')
         # put_cclasstable('classcodes.json')
-
